requirements.txt
langgraph
langchain
langchain-ollama
ipykernel
python-dotenv
streamlit
fastapi
uvicorn